It's working and the code is in it's discrete classes

+++ include/Timer.hpp
@@ -4,15 +4,15 @@
 class Timer
 {
     public:
-    Timer(double t);
-    void printTimer();
+        Timer(double t);
+        void printTimer();
     private:
-    double time;
-    int minutes;
-    int seconds;
-    bool run;
-    int counter;
-    double log;
+        double time;
+        int minutes;
+        int seconds;
+        bool run;
+        double counter;
+        double log;
 };
 
 #endif // TIMER_HPP
